Notes: Split Note into stateful component over NoteCard compound
